Add removers to MeshTerm
Added ConceptsEntity trait and applied it to Descriptor, SemanticType
and Term.
