+++ src/Ilios/CoreBundle/Entity/MeshSemanticType.php
@@ -5,6 +5,7 @@ namespace Ilios\CoreBundle\Entity;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\Mapping as ORM;
+use Ilios\CoreBundle\Traits\ConceptsEntity;
 use JMS\Serializer\Annotation as JMS;
 use Symfony\Component\Validator\Constraints as Assert;
 
@@ -30,6 +31,7 @@ class MeshSemanticType implements MeshSemanticTypeInterface
     use NameableEntity;
     use StringableIdEntity;
     use TimestampableEntity;
+    use ConceptsEntity;
 
     /**
      * @var string
@@ -104,45 +106,5 @@ class MeshSemanticType implements MeshSemanticTypeInterface
         $this->createdAt = new \DateTime();
         $this->updatedAt = new \DateTime();
         $this->concepts = new ArrayCollection();
-    }
-
-    /**
-     * @param Collection $concepts
-     */
-    public function setConcepts(Collection $concepts)
-    {
-        $this->concepts = new ArrayCollection();
-
-        foreach ($concepts as $concept) {
-            $this->addConcept($concept);
-        }
-    }
-
-    /**
-     * @param MeshConceptInterface $concept
-     */
-    public function addConcept(MeshConceptInterface $concept)
-    {
-        if (!$this->concepts->contains($concept)) {
-            $this->concepts->add($concept);
-            $concept->addSemanticType($this);
-        }
-    }
-
-    /**
-     * @param MeshConceptInterface $concept
-     */
-    public function removeConcept(MeshConceptInterface $concept)
-    {
-        $this->concepts->removeElement($concept);
-        $concept->removeSemanticType($this);
-    }
-
-    /**
-     * @return ArrayCollection|MeshConceptInterface[]
-     */
-    public function getConcepts()
-    {
-        return $this->concepts;
     }
 }
