+++ src/Ilios/CoreBundle/Entity/MeshSemanticTypeInterface.php
@@ -3,6 +3,7 @@
 namespace Ilios\CoreBundle\Entity;
 
 use Doctrine\Common\Collections\Collection;
+use Ilios\CoreBundle\Traits\ConceptsEntityInterface;
 use Ilios\CoreBundle\Traits\NameableEntityInterface;
 use Ilios\CoreBundle\Traits\TimestampableEntityInterface;
 use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
@@ -15,25 +16,7 @@ use Ilios\CoreBundle\Traits\IdentifiableEntityInterface;
 interface MeshSemanticTypeInterface extends
     IdentifiableEntityInterface,
     NameableEntityInterface,
-    TimestampableEntityInterface
+    TimestampableEntityInterface,
+    ConceptsEntityInterface
 {
-    /**
-     * @param Collection $concepts
-     */
-    public function setConcepts(Collection $concepts);
-
-    /**
-     * @param MeshConceptInterface $concept
-     */
-    public function addConcept(MeshConceptInterface $concept);
-
-    /**
-     * @param MeshConceptInterface $concept
-     */
-    public function removeConcept(MeshConceptInterface $concept);
-
-    /**
-     * @return ArrayCollection|MeshConceptInterface[]
-     */
-    public function getConcepts();
 }
