+++ src/Ilios/CoreBundle/Traits/ConceptsEntityInterface.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace Ilios\CoreBundle\Traits;
+
+use Doctrine\Common\Collections\ArrayCollection;
+use Doctrine\Common\Collections\Collection;
+
+use Ilios\CoreBundle\Entity\MeshConceptInterface;
+
+/**
+ * Interface ConceptsEntityInterface
+ * @package Ilios\CoreBundle\Traits
+ */
+interface ConceptsEntityInterface
+{
+    /**
+     * @param Collection $concepts
+     */
+    public function setConcepts(Collection $concepts);
+
+    /**
+     * @param MeshConceptInterface $concept
+     */
+    public function addConcept(MeshConceptInterface $concept);
+
+    /**
+     * @param MeshConceptInterface $concept
+     */
+    public function removeConcept(MeshConceptInterface $concept);
+
+    /**
+    * @return MeshConceptInterface[]|ArrayCollection
+    */
+    public function getConcepts();
+}
